PRSD-NONE: Renames GroupedJourneyExtensions to JourneyExtensions
