+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/helpers/extensions/journeyExtensions/JourneyExtensions.kt
@@ -2,7 +2,7 @@ package uk.gov.communities.prsdb.webapp.helpers.extensions.journeyExtensions
 
 import uk.gov.communities.prsdb.webapp.constants.BACK_URL_ATTR_NAME
 
-class GroupedJourneyExtensions {
+class JourneyExtensions {
     companion object {
         fun Map<String, Any>.withBackUrlIfNotNullAndNotCheckingAnswers(
             backUrl: String?,
